3.6

+++ src/main/java/it/algos/wiki24/backend/packages/parametri/ParList.java
@@ -4,6 +4,7 @@ import ch.carnet.kasparscherrer.*;
 import com.vaadin.flow.component.orderedlayout.*;
 import com.vaadin.flow.component.textfield.*;
 import static it.algos.vbase.backend.boot.BaseCost.*;
+import it.algos.vbase.ui.wrapper.*;
 import static it.algos.wiki24.backend.boot.WikiCost.*;
 import it.algos.wiki24.backend.list.*;
 import it.algos.wiki24.backend.logic.*;
@@ -20,16 +21,54 @@ public abstract class ParList extends WikiList {
 
     protected TextField searchGrezzo = new TextField();
 
-    protected TextField searchElaborato = new TextField();
+    protected TextField searchValido = new TextField();
 
-    protected IndeterminateCheckbox boxGrezzoVuoto;
+    protected IndeterminateCheckbox boxGrezzoPieno;
 
-    protected IndeterminateCheckbox boxElaboratoVuoto;
+    protected IndeterminateCheckbox boxValidoPieno;
 
     protected IndeterminateCheckbox boxUguale;
 
     public ParList(final WikiModulo crudModulo) {
         super(crudModulo);
+    }
+
+
+    @Override
+    protected void fixPreferenze() {
+        super.fixPreferenze();
+
+        this.usaBottoneDeleteAll = false;
+        this.usaBottoneResetDelete = false;
+        this.usaBottoneResetAdd = false;
+        this.usaBottoneDownload = false;
+        this.usaBottoneTransfer = false;
+        this.usaBottoneResetEntity = true;
+        this.usaBottoneWikiView = false;
+        this.usaBottoneWikiEdit = false;
+        this.usaBottoneWikiCrono = false;
+
+        super.usaInfoElabora = true;
+
+        this.usaBottoneSearch = false;
+        this.usaSearchPageId = false;
+        this.usaSearchWikiTitle = true;
+        this.usaBottoneExport = false;
+    }
+
+    protected void fixHeader() {
+        String message;
+        headerPlaceHolder.removeAll();
+
+        //Prima riga (infoScopo): Verde, bold, normale. Informazioni base: tavola (download) oppure Lista (upload) <br>
+        message = String.format("%s%sCostruita da BioMongo", TEXT_BASE, SPAZIO);
+        headerPlaceHolder.add(ASpan.text(message).bold().verde());
+
+        //Secondo gruppo: Blue, normale, normale. Logica di creazione/funzionamento della tavola <br>
+        message = String.format("Valore grezzo del parametro%sValore valido del parametro", FORWARD);
+        headerPlaceHolder.add(ASpan.text(message).blue());
+
+        super.fixHeader();
     }
 
     /**
@@ -38,7 +77,9 @@ public abstract class ParList extends WikiList {
     protected void fixTop() {
         wikiTopPlaceHolder.removeAll();
 
-        wikiTopPlaceHolder.deleteAll();
+        if (usaBottoneDeleteAll) {
+            wikiTopPlaceHolder.deleteAll();
+        }
         wikiTopPlaceHolder.elabora();
 
         if (usaBottoneEdit) {
@@ -53,51 +94,17 @@ public abstract class ParList extends WikiList {
         searchGrezzo.setClearButtonVisible(true);
         searchGrezzo.addValueChangeListener(event -> sincroFiltri());
 
-        searchElaborato.setPlaceholder(TAG_ALTRE_BY + FIELD_NAME_ELABORATO);
-        searchElaborato.getElement().setProperty("title", "Search: ricerca testuale da inizio del campo " + FIELD_NAME_ELABORATO);
-        searchElaborato.setClearButtonVisible(true);
-        searchElaborato.addValueChangeListener(event -> sincroFiltri());
+        searchValido.setPlaceholder(TAG_ALTRE_BY + FIELD_NAME_VALIDO);
+        searchValido.getElement().setProperty("title", "Search: ricerca testuale da inizio del campo " + FIELD_NAME_VALIDO);
+        searchValido.setClearButtonVisible(true);
+        searchValido.addValueChangeListener(event -> sincroFiltri());
 
         wikiTopPlaceHolder.build();
-        boxGrezzoVuoto = super.creaFiltroCheckBox(boxGrezzoVuoto, "Grezzo");
-        boxElaboratoVuoto = super.creaFiltroCheckBox(boxElaboratoVuoto, "Elaborato");
-        boxUguale = super.creaFiltroCheckBox(boxUguale, "Uguale");
-    }
-
-    /**
-     * Può essere sovrascritto, invocando PRIMA il metodo della superclasse <br>
-     */
-    protected void fixTop2() {
-        wikiTopPlaceHolder.removeAll();
-
-        boxGrezzoVuoto = super.creaFiltroCheckBox(boxGrezzoVuoto, "Grezzo");
-        //        boxGrezzoVuoto = new IndeterminateCheckbox();
-        //        boxGrezzoVuoto.setLabel("Grezzo");
-        //        boxGrezzoVuoto.setIndeterminate(false);
-        //        boxGrezzoVuoto.setValue(false);
-        //        boxGrezzoVuoto.addValueChangeListener(event -> sincroFiltri());
-        //        HorizontalLayout layout1 = new HorizontalLayout(boxGrezzoVuoto);
-        //        layout1.setAlignItems(Alignment.CENTER);
-
-        boxElaboratoVuoto = super.creaFiltroCheckBox(boxElaboratoVuoto, "Elaborato");
-        //        boxElaboratoVuoto = new IndeterminateCheckbox();
-        //        boxElaboratoVuoto.setLabel("Elaborato");
-        //        boxElaboratoVuoto.setIndeterminate(false);
-        //        boxElaboratoVuoto.setValue(false);
-        //        boxElaboratoVuoto.addValueChangeListener(event -> sincroFiltri());
-        //        HorizontalLayout layout2 = new HorizontalLayout(boxElaboratoVuoto);
-        //        layout2.setAlignItems(Alignment.CENTER);
-
-        boxUguale = super.creaFiltroCheckBox(boxUguale, "Uguale");
-        //        boxUguale = new IndeterminateCheckbox();
-        //        boxUguale.setLabel("Uguale");
-        //        boxUguale.setIndeterminate(false);
-        //        boxUguale.setValue(false);
-        //        boxUguale.addValueChangeListener(event -> sincroFiltri());
-        //        HorizontalLayout layout3 = new HorizontalLayout(boxUguale);
-        //        layout3.setAlignItems(Alignment.CENTER);
-
-        //        headerPlaceHolder.add(new HorizontalLayout(searchGrezzo, searchElaborato, layout1, layout2, layout3));
+        wikiTopPlaceHolder.add(searchGrezzo);
+        wikiTopPlaceHolder.add(searchValido);
+        boxGrezzoPieno = super.creaFiltroCheckBox(boxGrezzoPieno, FIELD_NAME_GREZZO);
+        boxValidoPieno = super.creaFiltroCheckBox(boxValidoPieno, FIELD_NAME_VALIDO);
+        boxUguale = super.creaFiltroCheckBox(boxUguale, FIELD_NAME_UGUALE);
     }
 
 
@@ -117,42 +124,21 @@ public abstract class ParList extends WikiList {
             }
         }
 
-        if (searchElaborato != null) {
-            String elaborato = searchElaborato.getValue();
-            if (textService.isValid(elaborato)) {
-                filtri.inizio(FIELD_NAME_ELABORATO, elaborato);
-                filtri.sort(Sort.by(Sort.Direction.ASC, FIELD_NAME_ELABORATO));
+        if (searchValido != null) {
+            String valido = searchValido.getValue();
+            if (textService.isValid(valido)) {
+                filtri.inizio(FIELD_NAME_VALIDO, valido);
+                filtri.sort(Sort.by(Sort.Direction.ASC, FIELD_NAME_VALIDO));
             }
             else {
-                filtri.remove(FIELD_NAME_ELABORATO);
+                filtri.remove(FIELD_NAME_VALIDO);
                 filtri.sort(basicSort);
             }
         }
 
-        super.fixFiltroCheckBox(boxGrezzoVuoto, FIELD_NAME_GREZZO_VUOTO);
-        //        if (boxGrezzoVuoto != null && !boxGrezzoVuoto.isIndeterminate()) {
-        //            filtri.uguale(FIELD_NAME_GREZZO_VUOTO, boxGrezzoVuoto.getValue());
-        //        }
-        //        else {
-        //            filtri.remove(FIELD_NAME_GREZZO_VUOTO);
-        //        }
-
-        super.fixFiltroCheckBox(boxElaboratoVuoto, FIELD_NAME_ELABORATO_VUOTO);
-        //        if (boxElaboratoVuoto != null && !boxElaboratoVuoto.isIndeterminate()) {
-        //            filtri.uguale(FIELD_NAME_ELABORATO_VUOTO, boxElaboratoVuoto.getValue());
-        //        }
-        //        else {
-        //            filtri.remove(FIELD_NAME_ELABORATO_VUOTO);
-        //        }
-
+        super.fixFiltroCheckBox(boxGrezzoPieno, FIELD_NAME_GREZZO_PIENO);
+        super.fixFiltroCheckBox(boxValidoPieno, FIELD_NAME_VALIDO_PIENO);
         super.fixFiltroCheckBox(boxUguale, FIELD_NAME_UGUALE);
-        //        if (boxUguale != null && !boxUguale.isIndeterminate()) {
-        //            filtri.uguale(FIELD_NAME_UGUALE, boxUguale.getValue());
-        //        }
-        //        else {
-        //            filtri.remove(FIELD_NAME_UGUALE);
-        //        }
-
     }
 
 }
